# Optimize the immutable STRING/BYTES dictionary lookup