@Jackie-Jiang 2022-06-10
It does not work for all characters because of the surrogate. Comparing UTF-8 byte-by-byte returns the code point order, but not the lexicographic order (expected). The code point order is the same as lexicographic order for code point U+0000 to U+D7FF, so we can check the values when building the string dictionary, and store a flag if all the characters are at this range.