@Jackie-Jiang 2022-06-10
For reference: https://en.wikipedia.org/wiki/UTF-16