@richardstartin 2022-06-10
> Is the UTF-8 normalized if it is always encoded from a String object?

yes, but not if it came from bytes (not sure if this is possible)