@richardstartin 2022-06-14
> Re normalization. Yes, you can do this, and if so then something like Unicode Normalization Form C (NFC) would be good. But then you need to make sure you also normalize all query text in the same way, of course. Plus there's the backwards compatibility issue for existing segments.
> 
> And it's not really an issue for this optimization, if you assume the user is ingesting text normalized (or not) as they see fit, and the query text uses the same normalization (or not). If you do decide to always normalize, then I'd suggest making that a separate PR.

Yes, there was a mistaken assumption when this topic was introduced that having a string object implied normalisation. It does not, so it’s an orthogonal issue to what this PR is solving.