@Jackie-Jiang 2023-01-21
The idea is re-implemented in #10044 and #10161